Removed duplicate calls to setViewMode()
Those duplicates are not needed. We can simply check if passed url count
is more than one. It presents problem when for example only one file out
of many exists - but it's unlikely to happen in practice and can be only
caused in normal circumstances by manual use of klook from console